added pagination and item, this has been moved from the original page
